Allow userspace access to the rolling stats
Use like: kill -SIGUSR1 6775; head -100 /dev/shm/dprfs-rolling-stats-*

So far we only track deletes and renames with the current change
given under "Now: ". It, and other counts, are all given in hex.
Those other counts do not appear at all if they happened more than
two changes in the past. Historic data is reliable as it won't be
changed any more. The current data should be considered incomplete
as the elements can change at any moment.
That said, it can probably be assumed any element before any latest
non-zero count is set. That is, in "0, 8, 12, 0" the first zero is
definitive as there's a later non-zero count. The same is true of
eight, however we don't know if twelve will shortly go to thirteen,
or if its set and it's the last zero which will go to one.