Change how dprfs handles mkdir
mkdir(DIR) now creates DIR-<timestamp> which is quickly renamed to
DIR. This to resolve a problem whereby a directory that is recreated having
been previously renamed/deleted would seem to have the contents of the
previous incarnation. Those contents persisting in that original dir
regardless of name used on gdrive keeps them separate from a later
recreation of the original dir.

Add test021.sh to test the above
Update previous tests with this change